pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-sniper"
version = "1.0.0"
description = "CLI tool that automates cold email outreach to recruiters and engineering managers"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "pyperclip>=1.8.0",
    "python-dotenv>=1.0.0",
    "prompt_toolkit>=3.0.0",
]

[project.scripts]
sniper = "job_sniper.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
